+++ NEWS.md
@@ -1,5 +1,9 @@
 # eDITH 0.2.0.9000
 
+## Major changes
+
+- `run_eDITH_optim`: argument `n.restarts` is added.
+
 ## Minor changes
 
 - `run_eDITH_optim`: `attempts.stats` is exported.